fix: window AVG/STDDEV promote precision (DECIMAL scale+4, INTEGER→DOUBLE)
WindowFunctions.computeOver returns double for AVG(INTEGER) instead of
truncating via integer division. MiniDbProject.buildWindowBatch creates
window column vectors with promoted scale for AVG/STDDEV/VAR (DECIMAL
scale+4, INTEGER→DOUBLE), matching the Aggregate path's precision
upgrades.

query47/53/57 still differ due to Calcite rewriting AVG(SUM) window
into a SUM window (semantic plan difference, not precision loss).

+++ minidb-server/src/main/java/com/minidb/server/plan/physical/WindowFunctions.java
@@ -210,7 +210,8 @@ public final class WindowFunctions {
                 if (isDecimal) {
                     return decimalSum.divide(BigDecimal.valueOf(count), java.math.MathContext.DECIMAL128);
                 }
-                return isFloating ? doubleSum / count : longSum / count;
+                // AVG(整数) 返回 double,避免整数除法截断
+                return isFloating ? doubleSum / count : (double) longSum / count;
             case MIN:
             case MAX:
                 return bestValue;
